+++ .yarnrc.yml
@@ -1,4 +1,6 @@
-defaultSemverRangePrefix: ''
+compressionLevel: mixed
+
+defaultSemverRangePrefix: ""
 
 enableGlobalCache: true
