+++ packages/core/admin/admin/modules.d.ts
@@ -1,0 +1,5 @@
+declare module '*.svg' {
+  const content: any;
+  // eslint-disable-next-line import/no-default-export
+  export default content;
+}
