--- packages/core/admin/admin/src/pages/MarketplacePage/components/NpmPackagesFilters/FilterSelect.js
@@ -1,43 +1,0 @@
-import React from 'react';
-
-import { Option, Select } from '@strapi/design-system';
-import PropTypes from 'prop-types';
-
-const FilterSelect = ({ message, value, onChange, possibleFilters, onClear, customizeContent }) => {
-  const computeFilterMessage = (filterName, count) => {
-    return `${filterName} (${count})`;
-  };
-
-  return (
-    <Select
-      data-testid={`${message}-button`}
-      label={message}
-      placeholder={message}
-      size="M"
-      onChange={onChange}
-      onClear={onClear}
-      value={value}
-      customizeContent={customizeContent}
-      multi
-    >
-      {Object.entries(possibleFilters).map(([filterName, count]) => {
-        return (
-          <Option data-testid={`${filterName}-${count}`} key={filterName} value={filterName}>
-            {computeFilterMessage(filterName, count)}
-          </Option>
-        );
-      })}
-    </Select>
-  );
-};
-
-FilterSelect.propTypes = {
-  message: PropTypes.string.isRequired,
-  value: PropTypes.array.isRequired,
-  onChange: PropTypes.func.isRequired,
-  possibleFilters: PropTypes.object.isRequired,
-  onClear: PropTypes.func.isRequired,
-  customizeContent: PropTypes.func.isRequired,
-};
-
-export default FilterSelect;
