--- packages/core/admin/admin/src/pages/MarketplacePage/components/NpmPackagesPagination/index.js
@@ -1,27 +1,0 @@
-import React from 'react';
-
-import { Box, Flex } from '@strapi/design-system';
-import { PageSizeURLQuery, PaginationURLQuery } from '@strapi/helper-plugin';
-import PropTypes from 'prop-types';
-
-const NpmPackagesPagination = ({ pagination }) => {
-  return (
-    <Box paddingTop={4}>
-      <Flex alignItems="flex-end" justifyContent="space-between">
-        <PageSizeURLQuery options={['12', '24', '50', '100']} defaultValue="24" />
-        <PaginationURLQuery pagination={pagination} />
-      </Flex>
-    </Box>
-  );
-};
-
-NpmPackagesPagination.propTypes = {
-  pagination: PropTypes.shape({
-    page: PropTypes.number.isRequired,
-    pageCount: PropTypes.number.isRequired,
-    pageSize: PropTypes.number.isRequired,
-    total: PropTypes.number.isRequired,
-  }).isRequired,
-};
-
-export default NpmPackagesPagination;
