--- packages/core/admin/admin/src/pages/MarketplacePage/constants.js
@@ -1,1 +1,0 @@
-export const MARKETPLACE_API_URL = 'https://market-api.strapi.io';
