+++ packages/core/admin/admin/tsconfig.json
@@ -6,6 +6,6 @@
       "@tests/*": ["./tests/*"]
     }
   },
-  "include": ["src", "tests", "custom.d.ts"],
+  "include": ["src", "tests", "custom.d.ts", "modules.d.ts"],
   "exclude": ["node_modules"]
 }
