+++ packages/core/helper-plugin/src/features/Tracking.tsx
@@ -76,11 +76,12 @@ interface EventWithoutProperties {
     | 'didEditMediaLibraryConfig'
     | 'didEditNameOfContentType'
     | 'didGenerateGuidedTourApiTokens'
+    | 'didGoToMarketplace'
     | 'didLaunchGuidetour'
     | 'didMissMarketplacePlugin'
     | 'didNotCreateFirstAdmin'
     | 'didNotSaveComponent'
-    | 'didPluginLearnMode'
+    | 'didPluginLearnMore'
     | 'didSaveComponent'
     | 'didSaveContentType'
     | 'didSearch'
