+++ packages/plugins/graphql/package.json
@@ -41,7 +41,7 @@
     "koa-compose": "^4.1.0",
     "lodash": "4.17.21",
     "nexus": "1.3.0",
-    "pluralize": "^8.0.0"
+    "pluralize": "8.0.0"
   },
   "devDependencies": {
     "cross-env": "^7.0.3",
